requirements.txt
playwright
beautifulsoup4
openai
chromadb
python-dotenv
fastapi
uvicorn